Set up Checkstyle + SpotBugs static analysis; fix all violations to zero
- Add Checkstyle 10.17.0 and SpotBugs 6.0.26 to build.gradle; both run
  as part of `./gradlew check` with ignoreFailures=false
- Add config/checkstyle/checkstyle.xml (style + correctness rules)
- Add config/spotbugs/exclude.xml (Minecraft/Forge false-positive suppressions)
- Run expand_star_imports.py to eliminate all 64 star imports across the codebase
- Fix all Checkstyle violations: ModifierOrder, NeedBraces, MissingSwitchDefault,
  SimplifyBooleanExpression, UnusedImports, redundant imports, trailing whitespace
- Fix all SpotBugs violations: WMI_WRONG_MAP_ITERATOR (9 keySet→entrySet),
  DMI_RANDOM_USED_ONLY_ONCE (→ThreadLocalRandom), ICAST_IDIV_CAST_TO_DOUBLE (4),
  DM_NEXTINT_VIA_NEXTDOUBLE, NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE (6),
  EQ_DOESNT_OVERRIDE_EQUALS / EQ_COMPARETO_USE_OBJECT_EQUALS (3),
  CN_IMPLEMENTS_CLONE_BUT_NOT_CLONEABLE, FE_FLOATING_POINT_EQUALITY,
  LI_LAZY_INIT_UPDATE_STATIC, SF_SWITCH_FALLTHROUGH, SS_SHOULD_BE_STATIC,
  NP_NONNULL_RETURN_VIOLATION (suppressed — Void return is correct Java)
- `./gradlew checkstyleMain spotbugsMain` both pass with zero violations

+++ src/main/java/hellfirepvp/astralsorcery/client/resource/SpriteSheetResource.java
@@ -88,7 +88,7 @@ public class SpriteSheetResource extends AbstractRenderableTexture {
     /** UV offset for the given absolute frame timer. */
     public Tuple<Float, Float> getUVOffset(long frameTimer) {
         int frame = (int) (frameTimer % frameCount);
-        return new Tuple<>((frame % columns) * uPart, (frame / columns) * vPart);
+        return new Tuple<>((frame % columns) * uPart, ((float) frame / columns) * vPart);
     }
 
     /**
